FIX: 탠스택 쿼리 적용
